--- src/main/resources/template/hello.vm
@@ -1,1 +1,0 @@
-<h1>Hello World!</h1>
