+++ src/main/resources/templates/bootstrap-from-cdn-body.mustache
@@ -1,0 +1,12 @@
+<!-- This stuff goes at the end of the body.  y See: https://getbootstrap.com/docs/3.3/getting-started/ -->
+
+<!-- JQuery See: https://code.jquery.com/ -->
+
+<script
+  src="https://code.jquery.com/jquery-3.3.1.min.js"
+  integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8="
+  crossorigin="anonymous"></script>
+  
+<!-- Latest compiled and minified JavaScript -->
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
+
