+++ src/main/resources/templates/course.mustache
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+{{> head }}
+<body>
+ {{> nav }}
+ <h1>Course Menu</h1>
+ <p>
+ 	This is a page for displaying courses.
+ </p>	
+ {{> bootstrap-from-cdn-body }}
+</body>
+</html>
