+++ src/main/resources/templates/head.mustache
@@ -1,0 +1,6 @@
+<head>
+  <title>Sparkjava Mustache Template Example 02</title>
+
+  {{> bootstrap-from-cdn-head }}
+
+</head>
