+++ src/main/resources/templates/intro.mustache
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html>
+{{> head }}
+<body>
+ {{> nav }}
+ <!--<h1>Hello {{name}}</h1> -->
+ <h1>GoGaucho</h1>
+ <p>
+ 	GoGaucho is a schedule planner for students at UCSB.
+ </p>	
+ {{> bootstrap-from-cdn-body }}
+</body>
+</html>
