+++ src/main/resources/templates/profile.mustache
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+{{> head }}
+<body>
+ {{> nav }}
+ <h1>Profile</h1>
+ <p>
+ 	This is a page for displaying user profile.
+ </p>	
+ {{> bootstrap-from-cdn-body }}
+</body>
+</html>
